[Refactor ♻️](divine controller): split controller into interface and controller
